erasmusplanner/deployment/backend/requirements.txt
fastapi
uvicorn
crewai
pydantic
python-dotenv